licencias validacion eliminar

+++ src/Brasa/RecursoHumanoBundle/Controller/BaseEmpleadoController.php
@@ -187,7 +187,12 @@ class BaseEmpleadoController extends Controller
                     foreach ($arrSeleccionados AS $codigoLicencia) {
                         $arLicencia = new \Brasa\RecursoHumanoBundle\Entity\RhuLicencia();
                         $arLicencia = $em->getRepository('BrasaRecursoHumanoBundle:RhuLicencia')->find($codigoLicencia);
-                        $em->remove($arLicencia);
+                        if ($arLicencia->getEstadoCerrada()== 0){
+                            $em->remove($arLicencia);
+                        }else {
+                            echo "<script>alert('El registro no se puede eliminar, ha sido cerrado el proceso de la licencia.');</script>";
+                        }
+                        //$em->remove($arLicencia);
                     }
                     $em->flush();
                     return $this->redirect($this->generateUrl('brs_rhu_base_empleados_detalles', array('codigoEmpleado' => $codigoEmpleado)));
